+++ .hintrc
@@ -41,6 +41,7 @@
       {
         "label": "off"
       }
-    ]
+    ],
+    "highest-available-document-mode": "off"
   }
 }
