+++ src/js_pref/main.js
@@ -11,15 +11,13 @@ import '../scss/main.scss';
 import '../scss/course-panel.scss';
 import '../scss/timetable.scss';
 import '../scss/course-list.scss';
-import '../css/add-on.css';
 
 import localforage from 'localforage/dist/localforage';
 
 import './attacher';
 import './course-panel';
-import './course-list';
 import './timetable';
-import './add_on';
+import './course-list';
 import * as Utils from './utils';
 
 const lastUpdate = require('../../package.json')['lastUpdate'];
